PerchPlacement/src/requirements.txt
open3d~=0.9.0.0
numpy~=1.19.1
Shapely~=1.7.0
matplotlib~=3.3.0
pyswarms~=1.1.0
scipy~=1.5.2
setuptools~=47.3.1
opencv-python~=4.2.0.34
triangle~=20200424
plyfile~=0.7.2
scikit-learn~=0.23.1
PyYAML~=5.3.1
rospy~=1.14.6
rospkg~=1.2.7
pymesh2~=0.3
trimesh~=3.7.12
vedo~=2020.3.4
pyvista~=0.25.3
mpltools~=0.2.0
rosbag~=1.14.7
MeshLabXML~=2018.3
mapbox_earcut